refactor: Terms Detail Page
