+++ frontend-hybrid/yakal/lib/screens/Login/TermsDetail/style.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import '../../../utilities/style/color_styles.dart';
+
+class TermsDetailStyle {
+  static const title = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 16,
+    fontWeight: FontWeight.w700,
+    height: 1.4,
+  );
+
+  static const content = TextStyle(
+    color: ColorStyles.gray6,
+    fontSize: 15,
+    fontWeight: FontWeight.w500,
+    height: 1.8,
+  );
+}
